split the batch from the actual API.
